Adds the add, remove, show, reload features inside methods.js
