+++ .prettierignore
@@ -1,2 +1,3 @@
 src/index.js
 server.js
+src/methods.js
